+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 * Allow GA4 link tracker to track to multiple child classes ([PR #3835](https://github.com/alphagov/govuk_publishing_components/pull/3835))
 * Add GA4 link tracking to govspeak callout links ([PR #3843](https://github.com/alphagov/govuk_publishing_components/pull/3843))
 * Allow inputs to be exempt from GA4 form tracker [REDACT] code ([PR #3848](https://github.com/alphagov/govuk_publishing_components/pull/3848))
+* Hide our cookie banner when JS is disabled ([PR #3830](https://github.com/alphagov/govuk_publishing_components/pull/3830))
 
 ## 37.2.4